+++ CHANGELOG.md
@@ -12,6 +12,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 
 ### Fixed
 - starred items in a feed can prevent further scrolling
+- Feed ordering uses wrong values (#2846)
 
 # Releases
 ## [25.0.0-alpha12] - 2024-10-23