+++ CHANGELOG.md
@@ -10,6 +10,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 - Performance improvements on item list
 - Rework feed and global sort ordering
 - Add 'r' shortkey to refresh feed and item list
+- Show automatically new news items on VueJS implementation (#2502)
 
 ### Fixed
 - starred items in a feed can prevent further scrolling